+++ include/Root.hpp
@@ -15,7 +15,7 @@ namespace GUI
         void draw_triangle(DrawPoint A, DrawPoint B, DrawPoint C) override;
         void draw_text(int row, int column, char *, Color color) override;
         void draw_textShadow(int row, int column, char *, Color color);
-        void draw_translate(int x, int y) override;
+        void draw_translate(float x, float y) override;
 
     public:
         Root();
